+++ packages/create-pantheon-decoupled-kit/__tests__/index.test.ts
@@ -199,6 +199,7 @@ To see this list at any time, use the --help command.`);
 		process.argv = ['node', 'bin.js', 'gatsby-wp'];
 		const data = {
 			_: ['gatsby-wp'],
+			dkHealthCheckVersion: versions['decoupled-kit-health-check'],
 			eslintConfigVersion: versions['eslint'],
 			force: false,
 			gatsby: true,