+++ packages/create-pantheon-decoupled-kit/src/generators/gatsby-wp.generator.ts
@@ -40,6 +40,7 @@ export const gatsbyWp: DecoupledKitGenerator<GatsbyWPAnswers, GatsbyWPData> = {
 	],
 	data: {
 		gatsbyPnpmPlugin: pnpm,
+		dkHealthCheckVersion: versions['decoupled-kit-health-check'],
 		wordpressKitVersion: versions['wordpress-kit'],
 		otherConfigsVersion: versions['other'],
 		eslintConfigVersion: versions['eslint'],