+++ packages/decoupled-kit-health-check/__tests__/GatsbyWordPressHealthCheck.test.ts
@@ -1,7 +1,7 @@
 import { vi } from 'vitest';
-import { NextWordPressHealthCheck } from '../src/classes/NextWordPressHealthCheck';
+import { GatsbyWordPressHealthCheck } from '../src/classes/GatsbyWordPressHealthCheck';
 
-describe('NextWordPressHealthCheck', () => {
+describe('GatsbyWordPressHealthCheck', () => {
 	beforeEach((context) => {
 		vi.restoreAllMocks();
 		context.logSpy = vi.spyOn(console, 'log');
@@ -9,17 +9,16 @@ describe('NextWordPressHealthCheck', () => {
 		delete process.env['PANTHEON_CMS_ENDPOINT'];
 		delete process.env['WP_APPLICATION_USERNAME'];
 		delete process.env['WP_APPLICATION_SECRET'];
-		delete process.env['PREVIEW_SECRET'];
 	});
 	it('should pass for a valid WPGRAPHQL_URL and invalid auth', async ({
 		logSpy,
 	}) => {
 		process.env['WPGRAPHQL_URL'] = 'https://wordpress.test';
-		const HC = new NextWordPressHealthCheck({ env: process.env });
+		const HC = new GatsbyWordPressHealthCheck({ env: process.env });
 		await HC.validateEndpoint()
+			.then((hc) => hc.validateWPGatsbyPlugin())
 			.then((hc) => hc.validateMenu())
 			.then((hc) => hc.validateAuth())
-			.then((hc) => hc.validatePreview())
 			.catch(console.error);
 
 		const result = logSpy.mock.calls.map(([call]) => call);
@@ -29,28 +28,11 @@ describe('NextWordPressHealthCheck', () => {
 		logSpy,
 	}) => {
 		process.env['PANTHEON_CMS_ENDPOINT'] = 'wordpress.test';
-		const HC = new NextWordPressHealthCheck({ env: process.env });
+		const HC = new GatsbyWordPressHealthCheck({ env: process.env });
 		await HC.validateEndpoint()
+			.then((hc) => hc.validateWPGatsbyPlugin())
 			.then((hc) => hc.validateMenu())
 			.then((hc) => hc.validateAuth())
-			.then((hc) => hc.validatePreview())
-			.catch(console.error);
-
-		const result = logSpy.mock.calls.map(([call]) => call);
-		expect(result).toMatchSnapshot();
-	});
-	it('should pass for a valid backend and valid auth but no preview secret', async ({
-		logSpy,
-	}) => {
-		process.env['PANTHEON_CMS_ENDPOINT'] = 'wordpress.test';
-		process.env['WP_APPLICATION_USERNAME'] = 'a1b2c3-d4e5g6';
-		process.env['WP_APPLICATION_PASSWORD'] = 'mysecretsecret';
-
-		const HC = new NextWordPressHealthCheck({ env: process.env });
-		await HC.validateEndpoint()
-			.then((hc) => hc.validateMenu())
-			.then((hc) => hc.validateAuth())
-			.then((hc) => hc.validatePreview())
 			.catch(console.error);
 
 		const result = logSpy.mock.calls.map(([call]) => call);
@@ -60,13 +42,12 @@ describe('NextWordPressHealthCheck', () => {
 		process.env['PANTHEON_CMS_ENDPOINT'] = 'wordpress.test';
 		process.env['WP_APPLICATION_USERNAME'] = 'a1b2c3-d4e5g6';
 		process.env['WP_APPLICATION_PASSWORD'] = 'mysecretsecret';
-		process.env['PREVIEW_SECRET'] = 'testpreviewsecret!';
 
-		const HC = new NextWordPressHealthCheck({ env: process.env });
+		const HC = new GatsbyWordPressHealthCheck({ env: process.env });
 		await HC.validateEndpoint()
+			.then((hc) => hc.validateWPGatsbyPlugin())
 			.then((hc) => hc.validateMenu())
 			.then((hc) => hc.validateAuth())
-			.then((hc) => hc.validatePreview())
 			.catch(console.error);
 
 		const result = logSpy.mock.calls.map(([call]) => call);
@@ -77,11 +58,11 @@ describe('NextWordPressHealthCheck', () => {
 	it.fails(
 		'should show a helpful error message if no required env vars are set',
 		async () => {
-			const HC = new NextWordPressHealthCheck({ env: process.env });
+			const HC = new GatsbyWordPressHealthCheck({ env: process.env });
 			await HC.validateEndpoint()
+				.then((hc) => hc.validateWPGatsbyPlugin())
 				.then((hc) => hc.validateMenu())
 				.then((hc) => hc.validateAuth())
-				.then((hc) => hc.validatePreview())
 				.catch((err) => console.log(err.message));
 		},
 	);
@@ -89,11 +70,25 @@ describe('NextWordPressHealthCheck', () => {
 		logSpy,
 	}) => {
 		process.env['PANTHEON_CMS_ENDPOINT'] = 'invalid.wordpress.test';
-		const HC = new NextWordPressHealthCheck({ env: process.env });
+		const HC = new GatsbyWordPressHealthCheck({ env: process.env });
 		await HC.validateEndpoint()
+			.then((hc) => hc.validateWPGatsbyPlugin())
 			.then((hc) => hc.validateMenu())
 			.then((hc) => hc.validateAuth())
-			.then((hc) => hc.validatePreview())
+			.catch((err) => console.log(err.message));
+
+		const result = logSpy.mock.calls.map(([call]) => call);
+		expect(result).toMatchSnapshot();
+	});
+	it('should show a helpful error message if the wp-gatsby plugin is not enabled is invalid', async ({
+		logSpy,
+	}) => {
+		process.env['WPGRAPHQL_URL'] = 'https://wordpress.test.no-plugin';
+		const HC = new GatsbyWordPressHealthCheck({ env: process.env });
+		await HC.validateEndpoint()
+			.then((hc) => hc.validateWPGatsbyPlugin())
+			.then((hc) => hc.validateMenu())
+			.then((hc) => hc.validateAuth())
 			.catch((err) => console.log(err.message));
 
 		const result = logSpy.mock.calls.map(([call]) => call);